better scripts

+++ scripts/build.sh
@@ -4,6 +4,7 @@ echo "REGISTRY_URL=${REGISTRY_URL}"
 echo "REGISTRY_NAMESPACE=${REGISTRY_NAMESPACE}"
 echo "IMAGE_NAME=${IMAGE_NAME}"
 echo "BUILD_NUMBER=${BUILD_NUMBER}"
+echo "ARCHIVE_DIR=${ARCHIVE_DIR}"
 
 # Learn more about the available environment variables at:
 # https://console.bluemix.net/docs/services/ContinuousDelivery/pipeline_deploy_var.html#deliverypipeline_environment
@@ -19,7 +20,25 @@ else
     exit 1
 fi
 
-echo -e "Building container image"
+echo -e "Checking registry current plan and quota"
+bx cr plan
+bx cr quota
+echo "If needed, discard older images using: bx cr image-rm"
+
+echo -e "Checking registry namespace: ${REGISTRY_NAMESPACE}"
+ns=$(bx cr namespaces | grep ${REGISTRY_NAMESPACE})
+if [ -z $ns ]; then
+    echo "Registry namespace ${REGISTRY_NAMESPACE} not found, creating it."
+    bx cr namespace-add ${REGISTRY_NAMESPACE}
+    echo "Registry namespace ${REGISTRY_NAMESPACE} created."
+else 
+    echo "Registry namespace ${REGISTRY_NAMESPACE} found."
+fi
+
+echo -e "Existing images in registry"
+bx cr images
+
+echo -e "Building container image: ${IMAGE_NAME}:${BUILD_NUMBER}"
 set -x
 bx cr build -t $REGISTRY_URL/$REGISTRY_NAMESPACE/$IMAGE_NAME:$BUILD_NUMBER .
 set +x
@@ -36,9 +55,11 @@ if [ -f ./chart/${CHART_NAME}/values.yaml ]; then
     #Update Helm chart values.yml with image name and tag
     echo "UPDATING CHART VALUES:"
     sed -i "s~^\([[:blank:]]*\)repository:.*$~\1repository: ${REGISTRY_URL}/${IMAGE_NAME}~" ./chart/${CHART_NAME}/values.yaml
-    sed -i "s~^\([[:blank:]]*\)tag:.*$~\1tag: ${APPLICATION_VERSION}~" ./chart/${CHART_NAME}/values.yaml
+    sed -i "s~^\([[:blank:]]*\)tag:.*$~\1tag: ${BUILD_NUMBER}~" ./chart/${CHART_NAME}/values.yaml
     cat ./chart/${CHART_NAME}/values.yaml
-    cp -r ./chart/ $ARCHIVE_DIR/
+    if [ ! -d $ARCHIVE_DIR/chart/ ]; then # no need to copy if working in ./ already
+      cp -r ./chart/ $ARCHIVE_DIR/
+    fi
 else 
     echo -e "${red}Helm chart values for Kubernetes deployment (/chart/${CHART_NAME}/values.yaml) not found.${no_color}"
     exit 1
